[NO-JIRA] Refactor
