+++ server/api/actions/make-payment/load.js
@@ -5,16 +5,23 @@ import { UNAUTHORISED_READ } from 'helpers/constants';
 import stripePaymentsAllowedAttributes from './stripePaymentsAllowedAttributes';
 import stripeInstance from './stripe';
 import getPaymentAndBalance from './getPaymentAndBalance';
+import sendUnsentPaymentReceipts from './sendUnsentPaymentReceipts';
 
 export default function loadSingle(req) {
   const reqSecured = reqSecure(req, stripePaymentsAllowedAttributes);
   return new Promise((resolve, reject) => {
     getPaymentAndBalance(reqSecured.body.paymentId)
       .then(payment => {
-        resolve({
-          ...payment,
-          email: 'REDACTED',
-        });
+        sendUnsentPaymentReceipts(payment)
+          .then(() => {
+            resolve({
+              ...payment,
+              email: 'REDACTED',
+            });
+          })
+          .catch(err => {
+            reject(err);
+          });
       })
       .catch(err => {
         reject(err);
